% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hdtable.r
\name{renderhdtable}
\alias{renderhdtable}
\title{Renders the hdtable.}
\usage{
renderhdtable(expr, env = parent.frame(), quoted = FALSE, options = NULL,
  readOnly = TRUE)
}
\arguments{
\item{expr}{The computation that leads to an output}

\item{env}{The env in which}

\item{quoted}{To be decided}

\item{options}{Loads of options}

\item{readOnly}{A vector of TRUE/FALSE values to indicate which of the
  columns should be readonly.}
}
\description{
Renders the hdtable.
}

